Testes unitarios para controller e service

+++ src/main/java/br/com/conta/model/dto/autenticacao/DadosAutenticacaoDTO.java
@@ -5,7 +5,7 @@ import jakarta.validation.constraints.NotNull;
 
 public record DadosAutenticacaoDTO(
 		@NotNull
-		Integer numero,
+		Long numero,
 		@NotBlank
 		String senha) {
